feat!: remove `/note/` slug
